feat: install dependencies and move ai to renderer
