+++ src/renderer/src/env.ts
@@ -1,0 +1,7 @@
+import { z } from 'zod'
+
+const envSchema = z.object({
+  VITE_OPENAI_API_KEY: z.string()
+})
+
+export const env = envSchema.parse(process.env)
