+++ src/renderer/src/services/openai.ts
@@ -1,5 +1,5 @@
 import { createOpenAI } from '@ai-sdk/openai'
-import { env } from '../../stript-duck-test/src/renderer/src/env'
+import { env } from '../env'
 
 export const openai = createOpenAI({
   apiKey: env.VITE_OPENAI_API_KEY
